* Update MinGW32/-w64 for _BSD_SOURCE not GNU. * Cleanup Net-SNMP in win32/64, albeit no suitably linked net-snmp binaries to test.  MSCRT binaries don't work correctly. * Recast INVALID_SOCKET and SOCKET_ERROR to int's to avoid long to int conversion warnings. * Update async example module for PGM socket level parameter in Win32 builds. * Cast interface index to unsigned when displaying to avoid long/int conversion warnings. * Cast sizeof() to socklen_t to avoid signed/unsigned warnings.

+++ openpgm/pgm/include/impl/sockaddr.h
@@ -52,8 +52,8 @@ PGM_BEGIN_DECLS
 #	define pgm_sock_strerror(e)		strerror(e)
 #	define pgm_error_from_sock_errno	pgm_error_from_errno
 #else
-#	define PGM_INVALID_SOCKET		INVALID_SOCKET
-#	define PGM_SOCKET_ERROR			SOCKET_ERROR
+#	define PGM_INVALID_SOCKET		(int)INVALID_SOCKET
+#	define PGM_SOCKET_ERROR			(int)SOCKET_ERROR
 #	define pgm_closesocket			closesocket
 #	define pgm_sock_errno()			WSAGetLastError()
 #	define pgm_sock_strerror(e)		pgm_wsastrerror(e)
